requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
autoflake==1.4
backports.entry-points-selectable==1.1.1
black==21.11b1
cfgv==3.3.1
chardet==4.0.0
click==8.0.3
colorama==0.4.4
discord-py-interactions==4.0.0
distlib==0.3.4
environs==9.3.5
filelock==3.4.0
flake8==3.9.2
identify==2.4.0
idna==3.3
isort==5.9.3
marshmallow==3.14.1
mccabe==0.6.1
multidict==5.2.0
mypy-extensions==0.4.3
nodeenv==1.6.0
orjson==3.6.3
pathspec==0.9.0
platformdirs==2.4.0
pre-commit==2.16.0
pycodestyle==2.7.0
pyflakes==2.3.1
python-dotenv==0.19.2
PyYAML==6.0
regex==2021.11.10
six==1.16.0
toml==0.10.2
tomli==1.2.3
typing_extensions==4.0.1
virtualenv==20.10.0
yarl==1.7.2
